requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
tiktoken
sentence-transformers
faiss-cpu
pinecone-client
PyPDF2
python-docx
requests
python-jose[cryptography]
passlib[bcrypt]
aiofiles
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
httpx

